[CACH] Add --secrets-only and --ssm-only to ElastiCache seeding script
This change is needed because the SAM stack initializes ElastiCache with
the AUTH user/password from AWS Secrets Manager. Plus, one of the
ElastiCache SSM parameters is the ElastiCache master replica endpoint,
which we can only get after ElastiCache is created.

Therefore, we must first seed the secrets with --secrets-only,
then create the SAM stack, then seed the ElastiCache parameters with
--ssm-only.